Move Maya Pyblish actions to colorbleed.maya module
