+++ colorbleed/api.py
@@ -12,10 +12,7 @@ from .plugin import (
 
 # temporary fix, might
 from .action import (
-
     get_errored_instances_from_context,
-    SelectInvalidAction,
-    GenerateUUIDsOnInvalidAction,
     RepairAction,
     RepairContextAction
 )
@@ -30,7 +27,5 @@ all = [
     "ValidateMeshOrder",
     # action
     "get_errored_instances_from_context",
-    "SelectInvalidAction",
-    "GenerateUUIDsOnInvalidAction",
     "RepairAction"
 ]
