+++ colorbleed/plugins/maya/publish/validate_animation_content.py
@@ -1,5 +1,6 @@
 import pyblish.api
 import colorbleed.api
+import colorbleed.maya.action
 
 
 class ValidateAnimationContent(pyblish.api.InstancePlugin):
@@ -14,7 +15,7 @@ class ValidateAnimationContent(pyblish.api.InstancePlugin):
     hosts = ["maya"]
     families = ["colorbleed.animation"]
     label = "Animation Content"
-    actions = [colorbleed.api.SelectInvalidAction]
+    actions = [colorbleed.maya.action.SelectInvalidAction]
 
     @classmethod
     def get_invalid(cls, instance):
