+++ colorbleed/plugins/maya/publish/validate_instance_has_members.py
@@ -1,5 +1,6 @@
 import pyblish.api
 import colorbleed.api
+import colorbleed.maya.action
 
 
 class ValidateInstanceHasMembers(pyblish.api.InstancePlugin):
@@ -8,7 +9,7 @@ class ValidateInstanceHasMembers(pyblish.api.InstancePlugin):
     order = colorbleed.api.ValidateContentsOrder
     hosts = ["maya"]
     label = 'Instance has members'
-    actions = [colorbleed.api.SelectInvalidAction]
+    actions = [colorbleed.maya.action.SelectInvalidAction]
 
     @classmethod
     def get_invalid(cls, instance):
