+++ colorbleed/plugins/maya/publish/validate_mesh_non_manifold.py
@@ -2,6 +2,7 @@ from maya import cmds
 
 import pyblish.api
 import colorbleed.api
+import colorbleed.maya.action
 
 
 class ValidateMeshNonManifold(pyblish.api.Validator):
@@ -16,7 +17,7 @@ class ValidateMeshNonManifold(pyblish.api.Validator):
     hosts = ['maya']
     families = ['colorbleed.model']
     label = 'Mesh Non-Manifold Vertices/Edges'
-    actions = [colorbleed.api.SelectInvalidAction]
+    actions = [colorbleed.maya.action.SelectInvalidAction]
 
     @staticmethod
     def get_invalid(instance):
