+++ colorbleed/plugins/maya/publish/validate_mesh_shader_connections.py
@@ -2,6 +2,7 @@ from maya import cmds
 
 import pyblish.api
 import colorbleed.api
+import colorbleed.maya.action
 
 
 def pairs(iterable):
@@ -76,7 +77,7 @@ class ValidateMeshShaderConnections(pyblish.api.InstancePlugin):
     hosts = ['maya']
     families = ['colorbleed.model']
     label = "Mesh Shader Connections"
-    actions = [colorbleed.api.SelectInvalidAction,
+    actions = [colorbleed.maya.action.SelectInvalidAction,
                colorbleed.api.RepairAction]
 
     def process(self, instance):
