+++ colorbleed/plugins/maya/publish/validate_renderlayer_aovs.py
@@ -1,5 +1,6 @@
 import pyblish.api
 
+import colorbleed.maya.action
 from avalon import io
 import colorbleed.api
 
@@ -24,7 +25,7 @@ class ValidateRenderLayerAOVs(pyblish.api.InstancePlugin):
     label = "Render Passes / AOVs Are Registered"
     hosts = ["maya"]
     families = ["colorbleed.renderlayer"]
-    actions = [colorbleed.api.SelectInvalidAction]
+    actions = [colorbleed.maya.action.SelectInvalidAction]
 
     def process(self, instance):
         invalid = self.get_invalid(instance)
