+++ colorbleed/plugins/maya/publish/validate_setdress_namespaces.py
@@ -1,5 +1,6 @@
 import pyblish.api
 import colorbleed.api
+import colorbleed.maya.action
 
 
 class ValidateSetdressNamespaces(pyblish.api.InstancePlugin):
@@ -17,7 +18,7 @@ class ValidateSetdressNamespaces(pyblish.api.InstancePlugin):
     label = "Validate Setdress Namespaces"
     order = pyblish.api.ValidatorOrder
     families = ["colorbleed.setdress"]
-    actions = [colorbleed.api.SelectInvalidAction]
+    actions = [colorbleed.maya.action.SelectInvalidAction]
 
     def process(self, instance):
