+++ colorbleed/plugins/maya/publish/validate_transform_naming_suffix.py
@@ -2,7 +2,7 @@ from maya import cmds
 
 import pyblish.api
 import colorbleed.api
-
+import colorbleed.maya.action
 
 SUFFIX_NAMING_TABLE = {'mesh': ["_GEO", "_GES", "_GEP", "_OSD"],
                        'nurbsCurve': ["_CRV"],
@@ -38,7 +38,7 @@ class ValidateTransformNamingSuffix(pyblish.api.InstancePlugin):
     optional = True
     version = (0, 1, 0)
     label = 'Suffix Naming Conventions'
-    actions = [colorbleed.api.SelectInvalidAction]
+    actions = [colorbleed.maya.action.SelectInvalidAction]
 
     @staticmethod
     def is_valid_name(node_name, shape_type):
